+++ spec/specification/json_spec.rb
@@ -121,10 +121,17 @@ module Pod
         hash['subspecs'].should == [{
           'name' => 'GreenBanana',
           'source_files' => 'GreenBanana',
-        }, {
+        }]
+        hash['testspecs'].should == [{
           'name' => 'Tests',
           'test_type' => :unit,
         }]
+      end
+
+      it 'writes test type for test subspec in json' do
+        @spec.test_spec {}
+        hash = @spec.to_json
+        hash.should.include '"name":"Tests","test_type":"unit"'
       end
 
       it 'can be loaded from an hash' do
@@ -141,9 +148,40 @@ module Pod
         hash = {
           'name' => 'BananaLib',
           'version' => '1.0',
+          'subspecs' => [{ 'name' => 'GreenBanana', 'source_files' => 'GreenBanana' }],
+          'testspecs' => [{ 'name' => 'Tests', 'test_type' => :unit }],
+        }
+        result = Specification.from_hash(hash)
+        result.subspecs.count.should.equal 2
+        result.test_specs.count.should.equal 1
+        result.test_specs.first.test_specification?.should.be.true
+        result.test_specs.first.test_type.should.equal :unit
+      end
+
+      it 'can load test specification from 1.3.0 hash format' do
+        hash = {
+          'name' => 'BananaLib',
+          'version' => '1.0',
           'subspecs' => [{ 'name' => 'GreenBanana', 'source_files' => 'GreenBanana' }, { 'name' => 'Tests', 'test_type' => :unit }],
         }
         result = Specification.from_hash(hash)
+        result.subspecs.count.should.equal 2
+        result.test_specs.count.should.equal 1
+        result.test_specs.first.test_specification?.should.be.true
+        result.test_specs.first.test_type.should.equal :unit
+      end
+
+      it 'can load test specification from 1.3.0 JSON format' do
+        json = '{"subspecs": [{"name": "Tests","test_type": "unit","source_files": "Tests/**/*.{h,m}"}]}'
+        result = Specification.from_json(json)
+        result.test_specs.count.should.equal 1
+        result.test_specs.first.test_specification?.should.be.true
+        result.test_specs.first.test_type.should.equal :unit
+      end
+
+      it 'can load test specification from json' do
+        json = '{"testspecs": [{"name": "Tests","test_type": "unit","source_files": "Tests/**/*.{h,m}"}]}'
+        result = Specification.from_json(json)
         result.test_specs.count.should.equal 1
         result.test_specs.first.test_specification?.should.be.true
         result.test_specs.first.test_type.should.equal :unit